backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
django-extensions==3.2.3
scikit-learn==1.3.2
pandas==2.1.4
numpy>=1.24.3,<2.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0
scikit-image==0.21.0
scipy==1.11.4
joblib==1.3.2
drf-yasg==1.21.11
requests==2.31.0
feedparser==6.0.10
Flask==3.0.0
Flask-CORS==4.0.0
mosec>=0.9.0
msgspec>=0.18.0
torch>=2.0.0
torchvision>=0.15.0
# pCR 예측 모델용
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.43.0
gunicorn==21.2.0
cryptography==41.0.7
nibabel==5.2.0
SimpleITK==2.3.1
pydicom==2.4.4
msgpack>=1.0.0
opencv-python>=4.8.0
# MRI Models (MONAI for 3D medical imaging)
monai>=1.3.0
# DICOM SEG 처리 (새로운 세그멘테이션 파이프라인)
pydicom-seg==0.4.1
highdicom
# LoRA for fine-tuning (SwinUNETR)
peft>=0.6.0
# PDF generation
reportlab==4.0.7

# Messenger Chat Server Dependencies
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0
redis>=4.5.0
django-cors-headers>=4.0.0
python-dateutil>=2.8.0

